# backend/src/main/java/com/example/dto/ErrorResponse.java
package com.example.dto;

/**
 * Standard error response DTO.
 * <p>
 * Used by GlobalExceptionHandler to return structured JSON error responses.
 * Matches the ErrorResponse schema in the API contract.
 */
public class ErrorResponse {

    private String timestamp;

    private int status;

    private String error;

    private String message;

    private String path;

    public ErrorResponse() {
    }

    public ErrorResponse(String timestamp, int status, String error, String message, String path) {
        this.timestamp = timestamp;
        this.status = status;
        this.error = error;
        this.message = message;
        this.path = path;
    }

    public String getTimestamp() {
        return timestamp;
    }

    public int getStatus() {
        return status;
    }

    public String getError() {
        return error;
    }

    public String getMessage() {
        return message;
    }

    public String getPath() {
        return path;
    }
}